updated tests and added code coverage

+++ CoolBlue.Products.Application/Common/IProductDataIntegration.cs
@@ -5,6 +5,6 @@ namespace CoolBlue.Products.Application.Common
     public  interface IProductDataIntegration
     {
         ProductTypeViewModel GetProductTypeByProductAsync(int productId, CancellationToken cancellationToken);
-        float GetSalesPriceAsync(int productId, CancellationToken cancellationToken);
+        double GetSalesPriceAsync(int productId, CancellationToken cancellationToken);
     }
 }
